Chore:Fixed all function into a function folder
